requirements.txt
asgiref==3.3.1
autopep8==1.5.6
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
Django==3.1.7
django-cors-headers==3.7.0
django-rest-auth==0.9.5
djangorestframework==3.12.2
idna==2.10
msgpack==1.0.2
Pillow==8.1.2
py-algorand-sdk==1.4.1
pycodestyle==2.7.0
pycparser==2.20
pycryptodomex==3.10.1
PyNaCl==1.4.0
python-decouple==3.4
pytz==2021.1
qrcode==6.1
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
style==1.1.0
toml==0.10.2
update==0.0.1
urllib3==1.26.4
